+++ src/Billing/BillingRouter.js
@@ -14,18 +14,18 @@ router.get('/', (req, res) => {
 });
 
 // Stripe Charge Route
-router.post('/testcharge', (req, res) => {
-    const cost = 2500;
-    // Basic Charge -- WORKING
-    stripe.charges.create({
-        amount: cost,
-        currency: 'cad',
-        source: 'tok_amex',
-        description: 'Hello Stripe World'
-    }, function (err, charge) {
-        if (err) res.status(200).json({ status: 'error', error: err });
-        res.status(200).json({ status: 'charge complete', data: charge });
-    })
+// router.post('/testcharge', (req, res) => {
+//     const cost = 2500;
+//     // Basic Charge -- WORKING
+//     stripe.charges.create({
+//         amount: cost,
+//         currency: 'cad',
+//         source: 'tok_amex',
+//         description: 'Hello Stripe World'
+//     }, function (err, charge) {
+//         if (err) res.status(200).json({ status: 'error', error: err });
+//         res.status(200).json({ status: 'charge complete', data: charge });
+//     })
 
     // charge made with New Customer -- NOT WORKING
     // stripe.customers.create({
@@ -38,23 +38,23 @@ router.post('/testcharge', (req, res) => {
     //     customer: customer.id
     // }))
     // .then(charge => res.status(200).json({ status: 'success', }))
-});
+// });
 
 
 // Charge made with Async/Await
-// router.post('/testcharge', async (req, res) => {
-//     try {
-//         let status = await stripe.charges.create({
-//             amount: 2000,
-//             currency: "usd",
-//             description: "an aexample charge",
-//             source: req.body
-//         });
+router.post('/testcharge', async (req, res) => {
+    try {
+        let { status } = await stripe.charges.create({
+            amount: 2000,
+            currency: "usd",
+            description: "an aexample charge",
+            source: req.body
+        });
 
-//         res.json({status});
-//     } catch (err) {
-//         res.status(500).end();
-//     }
-// });
+        res.json({status});
+    } catch (err) {
+        res.status(500).end();
+    }
+});
 
 module.exports = router;
