--- src/lspmgt_frontend/src/Components/Payment/Payment.js
@@ -1,20 +1,0 @@
-import React, {Component} from 'react';
-import {Elements, StripeProvider} from 'react-stripe-elements';
-import CheckoutForm from './Checkout';
-
-class Payment extends Component {
-  render() {
-    return (
-      <StripeProvider apiKey="pk_test_LwL4RUtinpP3PXzYirX2jNfR">
-        <div className="example">
-          <h1>React Stripe Elements Example</h1>
-          <Elements>
-            <CheckoutForm />
-          </Elements>
-        </div>
-      </StripeProvider>
-    );
-  }
-}
-
-export default Payment;
